LayoutTests/http/tests/security/window-name-after-same-origin-sub-frame-navigation.html

<!DOCTYPE html>
<html>
    <script>
        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.waitUntilDone();
        }

        console.log("window.name should equal \"test\" after a same-origin sub frame navigation.");
    </script>
    <iframe src="resources/window-name-test.html?127.0.0.1"></iframe>
</html>
